moving executor package to common
leaving e2e directory for leftovr tests and e2e.go
